Move routes to iot routes
